requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
